wired up users and relationships

+++ task-list/api/main.py
@@ -1,10 +1,10 @@
 from fastapi import FastAPI, HTTPException, Depends, status
-from sqlalchemy.orm import Session
+from sqlalchemy.orm import Session, joinedload
 from sqlalchemy import or_
 import datetime
 from faker import Faker
 import schemas, models
-from sqlalchemy.orm import Load, joinedload
+from typing import List
 
 from db import get_db, engine, Base
 
@@ -19,33 +19,38 @@ def root():
     return "Task API is running"
 
 # Get all tasks
-@app.get("/tasks")
+@app.get("/tasks", response_model=List[schemas.TaskSchema])
 def get_tasks(db: Session = Depends(get_db), search: str = ''):
-    tasks = db.query(models.TaskEntity).filter(or_(
+    tasks = db.query(models.TaskEntity) \
+    .options(joinedload(models.TaskEntity.user)) \
+    .filter(or_(
         models.TaskEntity.title.contains(search),
         models.TaskEntity.description.contains(search))).all()
-    return {"status": "success", "task": tasks}
+
+    return tasks
 
 # Get single task by id
-@app.get("/tasks/{task_id}")
+@app.get("/tasks/{task_id}",response_model=schemas.TaskSchema)
 def get_task(task_id: int, db: Session = Depends(get_db)):
-    task = db.query(models.TaskEntity).filter(models.TaskEntity.id == task_id).first()
-    task.user
+    task = db.query(models.TaskEntity) \
+        .options(joinedload(models.TaskEntity.user)) \
+        .filter(models.TaskEntity.id == task_id).first()
+
     if not task: raise HTTPException(status_code=404, detail="Task not found")
-    return {"status": "success", "task": task}
+    return task
 
 # Create task
-@app.post("/tasks", status_code=status.HTTP_201_CREATED)
-def create_task(task: schemas.TaskSchema, db: Session = Depends(get_db)):
+@app.post("/tasks", response_model=schemas.TaskSchema)
+def create_task(task: schemas.TaskPostSchema, db: Session = Depends(get_db)):
     new_task = models.TaskEntity(**task.dict())
     db.add(new_task)
     db.commit()
     db.refresh(new_task)
-    return {"status": "success", "task": new_task}
+    return new_task
 
 # Update task
-@app.put("/tasks/{task_id}", status_code=status.HTTP_201_CREATED)
-def update_task(task_id: int, task: schemas.TaskSchema, db: Session = Depends(get_db)):
+@app.put("/tasks/{task_id}")
+def update_task(task_id: int, task: schemas.TaskUpdateSchema, db: Session = Depends(get_db)):
     existing_task = db.query(models.TaskEntity).filter(models.TaskEntity.id == task_id).first()
     if not existing_task: raise HTTPException(status_code=404, detail="Task not found")
 
@@ -54,7 +59,7 @@ def update_task(task_id: int, task: schemas.TaskSchema, db: Session = Depends(ge
 
     db.commit()
     db.refresh(existing_task)
-    return {"status": "success", "task": existing_task}
+    return existing_task
 
 # Delete task
 @app.delete("/tasks/{task_id}")
@@ -64,7 +69,7 @@ def delete_task(task_id: int, db: Session = Depends(get_db)):
 
     db.delete(task)
     db.commit()
-    return {"status": "success", "task": task}
+    return task
 
 ############
 ## USERS ##
@@ -73,31 +78,35 @@ def delete_task(task_id: int, db: Session = Depends(get_db)):
 # Get all users
 @app.get("/users")
 def get_users(db: Session = Depends(get_db), search: str = ''):
-    users = db.query(models.UserEntity).filter(or_(
+    users = db.query(models.UserEntity) \
+        .filter(or_(
         models.UserEntity.fname.contains(search),
         models.UserEntity.lname.contains(search))).all()
-    return {"status": "success", "user": users}
+    return users
 
 # Get single user by id
-@app.get("/users/{user_id}")
+@app.get("/users/{user_id}", response_model=schemas.UserSchema)
 def get_user(user_id: int, db: Session = Depends(get_db)):
     user = db.query(models.UserEntity).filter(models.UserEntity.id == user_id).first()
-    user.tasks
     if not user: raise HTTPException(status_code=404, detail="User not found")
-    return {"status": "success", "user": user}
+
+    # an example of how to populate a field that is not in the database as long as the pydantic model has the field
+    #return schemas.UserSchema(**user.__dict__, full_name=f"{user.fname} {user.lname}")
+    return user
+
 
 # Create user
 @app.post("/users", status_code=status.HTTP_201_CREATED)
-def create_user(user: schemas.UserSchema, db: Session = Depends(get_db)):
+def create_user(user: schemas.UserPostSchema, db: Session = Depends(get_db)):
     new_user = models.UserEntity(**user.dict())
     db.add(new_user)
     db.commit()
     db.refresh(new_user)
-    return {"status": "success", "user": new_user}
+    return new_user
 
 # Update user
 @app.put("/users/{user_id}", status_code=status.HTTP_201_CREATED)
-def update_user(user_id: int, user: schemas.UserSchema, db: Session = Depends(get_db)):
+def update_user(user_id: int, user: schemas.UserUpdateSchema, db: Session = Depends(get_db)):
     existing_user = db.query(models.UserEntity).filter(models.UserEntity.id == user_id).first()
     if not existing_user: raise HTTPException(status_code=404, detail="User not found")
 
@@ -106,17 +115,20 @@ def update_user(user_id: int, user: schemas.UserSchema, db: Session = Depends(ge
 
     db.commit()
     db.refresh(existing_user)
-    return {"status": "success", "user": existing_user}
+    return existing_user
 
 # Delete user
 @app.delete("/users/{user_id}")
 def delete_user(user_id: int, db: Session = Depends(get_db)):
+    tasks = db.query(models.TaskEntity).filter(models.TaskEntity.user_id == user_id).all()
+    if len(tasks) > 0: raise HTTPException(status_code=400, detail="User has tasks and cannot be deleted")
+
     user = db.query(models.UserEntity).filter(models.UserEntity.id == user_id).first()
     if not user: raise HTTPException(status_code=404, detail="User not found")
 
     db.delete(user)
     db.commit()
-    return {"status": "success", "user": user}
+    return user
 
 
 ################################
